[nick/#10] fix/뒤로가기 버튼 콜백 위치 변경
