+++ Nick/Nike/app/src/main/java/com/clone/nike/MainActivity.kt
@@ -22,9 +22,6 @@ class MainActivity : AppCompatActivity() {
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        //뒤로가기 버튼 인식 콜백
-        onBackPressedDispatcher.addCallback(this, onBackPressedCallback)
-
         //BottomNav (Navigation 기반)
         val navHostFragment =
             supportFragmentManager.findFragmentById(R.id.main_fragment_container)
@@ -76,20 +73,5 @@ class MainActivity : AppCompatActivity() {
 //                else -> false
 //            }
 //        }
-    }
-
-    // 뒤로가기 두번 클릭 시 종료
-    private val onBackPressedCallback = object : OnBackPressedCallback(true) {
-        var pressedTime:Long = 0;
-        override fun handleOnBackPressed() {
-            if(System.currentTimeMillis() - pressedTime >= 2000) {
-                pressedTime = System.currentTimeMillis()
-                Toast.makeText(this@MainActivity, "뒤로가기 버튼을 한번 더 누르면 종료됩니다.", Toast.LENGTH_SHORT).show()
-            }
-            else {
-                finishAffinity()
-            }
-        }
-
     }
 }
